Move ViewingRoomCarousel to Components folder
