+++ src/v2/Apps/ViewingRoom/Components/ViewingRoomsFeaturedRail.tsx
@@ -2,7 +2,7 @@ import React from "react"
 import { Box, Flex, Link, MediumCard, Sans, Spacer } from "@artsy/palette"
 import { ViewingRoomsFeaturedRail_featuredViewingRooms } from "v2/__generated__/ViewingRoomsFeaturedRail_featuredViewingRooms.graphql"
 import { createFragmentContainer, graphql } from "react-relay"
-import { ViewingRoomCarousel } from "../Routes/Works/Components/ViewingRoomCarousel"
+import { ViewingRoomCarousel } from "./ViewingRoomCarousel"
 import { getTagProps } from "../Components/ViewingRoomsLatestGrid"
 
 interface ViewingRoomsFeaturedRailProps {
